Move SoundManager to the Managers namespace
